Reorganize config
This commit rewrites and restructures
the configuration added in the previous commit.

It was principally undertaken as an exercise
in understanding the boilerplate config code
taken from Jeremy Evans' `roda-sequel-stack` application skeleton.

This version of the code is longer and not as DRY,
but is arguably more declarative.
Namely, it uses case statements to completely encapsulate
where the application logic differs in different environments.

It also enforces a convention for naming of constants,
where TitleCased is for classes and modules
while ALL_CAPS is for everything else.

Another goal of this restructuring was to begin organizing the repo
in a way that more closely resembles Rails conventions.
As a matter of personal preference,
I like to have a small set of top-level directories
that meaningfully reflect the architecture of the application.