bfx: Give search results a unique `key` attr
When rendering a list in React
(e.g., mapping a JS array to a collection of <li> elements),
each element must have a unique `key` attribute
that uniquely identifies it from one render to the next.
This helps React determine which elements need to be rerendered,
and is described in detail here:

https://reactjs.org/docs/lists-and-keys.html

~This commit adds an `id` property to military bases~
~to use as the `key` attribute when rendering a list of search results.~

=== Redaction

In the interest of data privacy,
this commit was retroactively modified
to remove the assets/data/sites.json file.
It contained a list of US military bases
and their corresponding lat/lng coordinates.

Now that this data has been removed,
this repository may be safely hosted on a public platform.