rfg: Utilize Mapbox GL's LngLatLike interface
=== Redaction

In the interest of data privacy,
this commit was retroactively modified
to remove the assets/data/sites.json file.
It contained a list of US military bases
and their corresponding lat/lng coordinates.

Now that this data has been removed,
this repository may be safely hosted on a public platform.