rfg: Extract deploymentsReducer as pure function
React has an undocumented quirk wherein
useReducer dispatch functions are sometimes called twice in a row.
I don't really understand it, but it appears to be benign, and
Dan Abramov says it's "expected behavior in some cases"[0].
Nevertheless, it can be "fixed" by moving the reducer function
outside the scope of the component it's used in.

In this case, the reducer function originally wasn't a pure function
(it had the side effect of setting the `prompt` state variable),
so to extract it, that line of code had to be moved elsewhere.

[0]: https://github.com/facebook/react/issues/16295#issuecomment-521630581