bfx: Complete "sad path" handling in combobox input
There are four "sad paths" that must be handled
when a user abandons the combobox with partial input:

1. user hits <Esc>,
2. user hits <Tab>/<S-Tab>,
3. user clicks out of the input field, and
4. user clears input (<Ctrl-A>+<Backspace>).

In all four cases, the input should be cleared
and the result list should be closed.
(Caveat: If a selection has already been made,
1 through 3 should reset the input to that value.)

Originally, this behavior was broken for sad paths #2-3
(the result list was not automatically closed)
because the author mistakenly assumed that
`setControlledInput` would implicitly trigger `onInputValueChange`.

This commit adds the necessary coverage for this edge case.