rfg: Merge `searchResults` state var into `prompt`
This commit expands the scope of the `prompt` state var,
which started out as just a flag for the Map component
to know when to present a date picker popup.

Now, the `prompt` state var encapsulates the current focal UI task,
carrying some state related to it.
This eliminates one pair of state/setState props
defined in the top-level App component
and passed down to almost all levels of the component tree,
and unifies some of the logic for how such state changes are handled.

It would be nice to unify the `focusedResult` state var this way too,
but all attempts thus far have involved increasing complexity,
not reducing it.