rfg: Convert combobox callbacks to effect hooks
This commit models all combobox UI behavior as React effect hooks.

A case could be made that this is a bad thing,
since it adds a layer of indirection to the control flow:
we've gone from

    event -> callback

to

    event -> set state -> render effect

But there are two practical benefits to this remodeling:

  1. callbacks can be broken up or combined into distinct actions
     (e.g., separate setControlledInput from debouncer.flush, or
     unify inputValue/isOpen callbacks triggered by <Esc>);
  2. callbacks now support cleanup logic
     (e.g., onHighlightedIndexChange can highlight map markers,
     and then _reset them on exit_).

Both of these advantages will be utilized in upcoming commits.