chr: Upgrade downshift package
This package upgrade violates semantic versioning--that is,
downshift saw a minor version change (5.0.3 -> 5.2.5)
that introduced breaking changes.
As a result, it required an extensive rewrite of existing logic.

=== Background

The onStateChange prop is a callback function invoked any time
user interaction causes the internal state of the combobox to change.
Each time it is called, it updates a React state variable
with (an object representing) the internal state of the combobox.
Changes to this state are monitored to trigger various effect hooks.

After the upgrade, the content of this internal state object changed:
where before, it always contained the same four properties,

  * isOpen
  * selectedItem
  * inputValue
  * highlightedIndex

in the latest version, it only contains the properties
that changed during the last action (plus a new action "type").
For instance, if the state update was triggered by keyboard input,
the internal state object might be:

    { type: '__input_change__', inputValue: 'foo' }

Inconveniently, this object format change
broke the triggering of some effect hooks, requiring a massive rewrite.
On the upside, it enabled a cleaner separation of concerns between them.