bfx: Apply workaround for mapboxgl-js bug
Version 1.8.0 of mapboxgl-js does not support a "padding" option
for the .flyTo() function.

Newer versions (up to 1.11.0-beta.1 as of this writing) fix this,
but come with a strange bug:
padding works perfectly initially,
but once a base has been selected for the first deployment,
all subsequent flyTo/fitBounds calls are off-center.

(My suspicion is that it's related to mapbox-gl-js #9477,[0]
in which padding is applied twice...?)

[0]: https://github.com/mapbox/mapbox-gl-js/issues/9477