dev: Upgrade pry
In bc72b10, we upgraded to an unreleased commit of pry (272b329[0])
because it fixed/silenced a warning about Binding#source_location.

This version worked fine for the most part,
but contained a bug that broke tab completion in the REPL.

This commit upgrades to the latest stable release,
which resolves the tab completion bug.

[0]: https://github.com/pry/pry/commit/272b3290b5250d28ee82a5ff65aa3b29b825e37b