ftr: Add guided tour
This initial version of the guided tour is not perfect,
but probably good enough.

It would have been preferable
to show a live demo/simulation of searching for a base--
from typing into the search box
to selecting a base from the search results list
to the resulting map refitting--
but apparently it would violate the browser's security model
to allow JS to manipulate form input on the user's behalf.

The only piece that's missing at this point
is a screenshot of the generated chart/report.
(It may be better to present that as an SVG or raw HTML,
if that's how the actual product is going to be presented.)