bfx: Replace buggy focusDebouncer with better collision detection
67892ae0 added a focusDebouncer() method to be used
"whenever multiple, near-simultaneous events want to setUIFocus."
The main use case for this was clicking on the delete-deployment button,
which triggered both removeHandler() (which reset UI focus)
and a general onFocus listener for the search unit
(which zoomed UI focus down to a given base).

This approach of letting both handlers run and then debouncing them
suffered from jittery performance on slow machines.
Rather than firing both and trying to debounce them,
this commit simply adds better detection logic
to determine which event should fire, and suppress the other.