dpl: Fix Procfile $PORT usage
Forego (presumably along with other Procfile utilities)
is designed to implicitly set PORT=5000 for all the processes it spawns.

Thus, rather than explicitly setting, say,

    web: bundle exec rackup -p 3000

you're supposed to

    web: bundle exec rackup -p $PORT

so that

    $ forego start            # PORT=5000, implicitly
    $ PORT=3000 forego start  # PORT=3000, explicitly

N.b. this is incompatible with Zammad's approach to Procfiles

    web: bundle exec rails server -p ${PORT:=3000}

since forego always implicitly defines $PORT,
so there's no way for the default value of 3000 to ever get used.