dep: Use Bundler.require
I had previously neglected this approach to configuration
(in Rails apps, it's done automatically in config/application.rb).

It's described in detail in the Bundler docs,[0] but here's the upshot:

    require 'rubygems'

    # put all gems on load path
    require 'bundler/setup'

    # put specified groups of gems on load path
    require 'bundler'
    Bundler.setup(:default, :ci)

    # require groups of gems
    Bundler.require(:default, :ci)

This commit adopts this pattern of requiring dependencies.
It also prevents sequel_pg from being required,
which would otherwise lead to a NameError.[1]

[0]: https://bundler.io/v2.0/guides/groups.html
[1]: https://github.com/jeremyevans/sequel/pull/720#issuecomment-570800356