ftr: Use strings for base IDs (not ints)
The CSV file originally used
to import VA base data into the frontend application
contained numerical ID values for each record
which were generated by the development team.

The client recently provided a sample SQL dataset
which used a string as the primary key on the bases table
(e.g., "VA1259").
~This commit updates the `bases.csv` file~
~used to populate the frontend application with VA base data~
~to match the string IDs in the client's sample dataset.~

=== Redaction

In the interest of data privacy,
this commit was retroactively modified
to remove the assets/data/sites.json file.
It contained a list of US military bases
and their corresponding lat/lng coordinates.

Now that this data has been removed,
this repository may be safely hosted on a public platform.