ftr: Add ExposureSummary UI component
This commit covers a lot of ground,
all in the service of a single UI change:
the "Submit" button does something now.

Many of the commits leading up to this one were written
to make this one possible.
If I'm being perfectly honest,
the complexity got a bit out-of-hand,
so this commit does more than one thing
and leaves a few of those things unfinished.
However, a dogmatic adherence to separating changes
into isolated commits with clearly-defined boundaries
was really standing in the way of getting the job done.

Some of the changes in this commit could use further refactoring and
DRYing up, such as the "exposure-menu" li styling.