rfg: Rename global state variable (uiFocus -> userFlow)
=== Background

If the application's user flow can be thought of
as a flow chart (or a state machine),
the uiFocus variable was intended to encapsulate
the user's current position in it.

To put it another way, the presentation of certain page elements
is a function of what the user is doing right now; e.g.,

* enter text into base search box => map markers over each result
* mouseover completed deployment => animate a floating map marker

These visual arrangements depend purely on the user's interaction,
meaning that each "node" in the user flow has its own layout.

(Note that "position in user flow" is distinct
from other kinds of application state:
whether you've added five deployments or none,
the "review deployments" layout should look a certain way.)

This uiFocus variable manages this relationship
by storing the user's current interaction with the page;
CSS classes and side effects are then applied based on its value.

=== Why the change?

This variable was originally created with a fluent API in mind:

    if (uiFocus.on === 'search results') { ... }

Then, c2e97a3 introduced the second half of the application,
which adds a whole new mode to the user interface.
This growing complexity deserved some namespacing
to keep these arbitrary "ui focus" targets explicit
and easier to understand.

Now, the userFlow variable takes the following format:

    {
      mode: ('deployment builder' | 'exposure data'),
      action: ...,
      <metadata>: ...,
    }

Maybe not perfect, but allows for much more flexibility than before.