dpl: Handle docker-compose startup quirks during app launch
When deploying the app via `docker-compose up`,
the postgres container ("db") needs a few extra seconds to go live
before the app container ("web") can successfully connect to it.

Previous versions of docker-compose used the `depends_on` option
to explicitly manage the startup order of docker-compose containers,
but the official recommendation now is to change your application
so that it explicitly and gracefully handles connection failures.[0]

This commit addresses connection failures during app launch,
but not during normal operation.

[0]: https://docs.docker.com/compose/startup-order/