uix: Make first tab-focus go to search input
78a94b4 introduced a regression in the UI
in which the user must hit Tab _twice_ after page load
for focus to go to the search input.

This was the result of changes in
`display` and `visibility` CSS rules on .search-panel__back-button,
which comes before the first search input in the DOM.

The fix was to re-apply the correct visibility rules on the button.
`transition-delay` rules had to be adjusted
to ensure the show/hide animation would not be cut short
by an immediate toggling of visibility.