ftr: Accept app state from URL fragment (hash) parameters
This commit is in preparation for two related, upcoming features:
1) browser history integration and 2) storable app state.

It supports these features by enabling the app to parse URL fragments
(a.k.a. "hash", as in `window.location.hash`):

    https://localhost:3000/#VA0518[]=2018-01-01,2018-02-01
                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

Now, visiting the URL above will launch the application
with one deployment (base + period) pre-selected.

(To learn more about URL fragments
and RFCs/specifications/conventions around them, see Wikipedia[0].)

In a forthcoming commit, changes to the app's state
will be pushed to the browser's history stack
so that paging Back and Forward will "walk through"
the history of the user's current session.
The browser's location bar will be updated in real time
with a URL fragment for each such "snapshot" of app state.
These updated URLs can then be bookmarked or copied-and-pasted
to recreate the app state in another session.

(Currently, this commit only adds the ability to _read_ such fragments
and load their contents into app state.)

[0]: https://en.wikipedia.org/wiki/URI_fragment