rfg: Prefer native Date() over moment()
=== Context

This application uses moment.js to store dates. Why?
Because that is the format consumed/provided by
the ant.design DatePicker component.[0]

Prior to this commit,
any time we encountered a date in a different format (string, Date()),
we converted it to moment() and then stored it to app state
(`deployments.period`).
This was convenient for compatibility with the DatePicker component,
but it was not strictly necessary.

=== Problems

I would have preferred native JS Date() objects in the first place,
if only to minimize dependencies and bundled code.
But recently, saving moment.js dates to app state (`deployments.period`)
presented another problem:
using moment.js breaks the app's browser history integration.

This app state needs to be periodically saved to `history.state`
so that it can be restored as the user pages back and forth
through browser/session history.
However, the browser serializes any such data
before storing it to `history.state`[1]
using the structured clone algorithm,[2]
which does not work with moment.js objects.

=== Fix

Instead of saving dates as moment() objects
and passing them directly to/from the DatePicker component,
we now store them as Date() objects
and convert them to/from moment() objects
before being passed to (or after being received from)
the DatePicker component.

[0]: https://ant.design/components/date-picker/
[1]: https://developer.mozilla.org/en-US/docs/Web/API/History_API/Working_with_the_History_API#The_replaceState_method
[2]: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm