bfx: Fix stale SearchUnit input field during history nav
Newly added browser history integration revealed a bug
in the SearchUnit component:

Its <input> value is a controlled state var
which can only be in one of two states:

  1. blurred: `deployment.base.name || ''`
  2. focused: whatever the user is currently typing

However, prior to this commit, there were edge cases
in which `deployment.base.name` was undefined
but the input field would show the most recent base name,
or the base name was defined but the input field would be empty.
(Namely, when `deployments` was modified under the hood;
i.e., via paging back and forward through browser history).

This commit keeps `controlledInput` in sync with `deployment.base.name`
through the use of a `useEffect` hook.