dev: Redact sites.json from repository
The purpose of this commit is to make this repository
safe to host on a public platform like GitHub.
It concludes an editing/rebasing of historical commits
to remove all traces of the `sites.json` file,
which contained a list of US military bases in the SADA region
along with lat/lng coordinates for each.

This involved the modification of four separate commits,
each of which touched (or created) the `sites.json` file:

* d171937 - ftr: Use strings for base IDs (not ints)
* cf0e9ab - rfg: Utilize Mapbox GL's LngLatLike interface
* a1e3e14 - bfx: Give search results a unique `key` attr
* 2dfa97c - Build map & search form skeleton

One consequence of this change is that
the contents of this repo are no longer sufficient
to compile a working version
of the front end (React SPA) of this application.
Now, you will need to obtain a separate copy of `sites.json`
saved to the `data/` directory.