bfx: Memoize retrieved exposure data when possible
Previously, the application would fire two separate GET requests
every single time the Submit button was clicked--
even if the selected deployments had not changed.

Since this application's data store is fixed
(i.e., it does not change on a day-to-day basis),
the retrieved data can be assumed to be identical
when selected deployments have not changed.
This commit adds logic to determine whether the last-retrieved data
is stale or not, and only fetch when it is.