uix: Show spinner on chart page while fetching API data
When clicking "Submit" two switch from map mode to chart mode,
it may take a second or two for the frontend
to fully retrieve exposure data from the backend
and render the results.

Originally, the chart page would be blank (or worse, show stale data)
during this interval.
Now, it shows a loading spinner instead.