bfx: Fix misconfigured chart tooltip
Prior to this commit,
the chart tooltip had the wrong options applied to it,
leading to failures whenever selecting multiple deployments
covering non-identical periods (timeframes).

Specifically, the tooltip would jump all over the chart
instead of tracking the mouse pointer position.
This problem was erroneously filed as an issue on chartjs/Chart.js:

https://github.com/chartjs/Chart.js/issues/8126