rfg: Store bases JSON metadata as hash, not array
When the frontend application needs access to bases metadata,
it reads from a pre-generated `sites.json` file rather than query the DB
(because it's faster than hitting the network).

Prior to this commit, the JSON metadata was a simple array of hashes:

    [
      {
        "id": "VA0001",
        "country": "Iraq",
        "name": "Camp Krusty",
        "lng": "0.33333e2",
        "lat": "0.44383e2",
      },
      ...
    ]

Now, it's a hash of hashes:

    {
      "VA0001": {
        "id": "VA0001",
        "country": "Iraq",
        "name": "Camp Krusty",
        "lng": "0.33333e2",
        "lat": "0.44383e2",
      },
      ...
    }

Why? Because we have to look up bases by their IDs in multiple places,
and it's computationally less expensive this way.