ftr: Show per-base exposure stats on summary page (2/3)
In this commit, we modify the GET /exposure_stats endpoint
to return averages & standard deviations
for not only the entire queried set of deployments,
but also for each individual base in the query.

These per-base averages and standard deviations
are required by the new ExposureSummary page,
which must display separate tables for this data.

On the frontend, the `exposureStats` object
had to be converted to a Map object because when we iterate over it,
we want to guarantee that the "aggregate" key
(i.e., stats for all deployments in the query)
is yielded first.